🎨 Edit Board-Add Function
Board : Event
