+++ src/main/java/com/phoenix/howabouttoday/board/controller/ContactController.java
@@ -1,0 +1,39 @@
+package com.phoenix.howabouttoday.board.controller;
+
+import com.phoenix.howabouttoday.board.dto.*;
+import com.phoenix.howabouttoday.board.service.BoardService;
+import com.phoenix.howabouttoday.board.service.EventService;
+import com.phoenix.howabouttoday.board.service.FAQService;
+import com.phoenix.howabouttoday.config.auth.LoginUser;
+import com.phoenix.howabouttoday.member.dto.SessionDTO;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import javax.validation.Valid;
+import java.util.List;
+
+@Controller
+@RequiredArgsConstructor
+public class ContactController {
+
+    // 고객센터
+    @GetMapping("contact")
+    public String getContact(@LoginUser SessionDTO sessionDTO, Model model){
+
+        if(sessionDTO != null) {
+            model.addAttribute("sessionDTO", sessionDTO);
+        }
+
+        return "board/contact";
+    }
+
+    @PostMapping("contact")
+    public String postContact(){
+        return "board/contact";
+    }
+
+}
