+++ src/main/java/com/phoenix/howabouttoday/board/dto/EventEditDTO.java
@@ -1,21 +1,15 @@
 package com.phoenix.howabouttoday.board.dto;
 
-import com.phoenix.howabouttoday.board.entity.Event;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @NoArgsConstructor
 @Getter
 @Setter
-public class EventAddDTO {
+public class EventEditDTO {
 
     // 게시판 작성 페이지 : Event
 
@@ -25,5 +19,15 @@ public class EventAddDTO {
     private String eventTitle; // 이벤트 게시글 제목
 
     private String date;
+//    private LocalDate eventStart; // 이벤트 시작일
+//    private LocalDate eventEnd; // 이벤트 종료일
+//    private List<EventImageDTO> eventImageList; // 이벤트 이미지
 
+//    public LocalDate getEventStart() {
+//        return LocalDate.parse(date.split("-")[0]);
+//    }
+//
+//    public LocalDate getEventEnd() {
+//        return LocalDate.parse(date.split("-")[1]);
+//    }
 }
