+++ src/main/resources/templates/board/event-add.html
@@ -80,7 +80,7 @@
                                 <div class="form-box event-box">
                                     <div class="form-content">
                                         <div class="file-upload-wrap file-upload-wrap-3">
-                                            <input type="file" th:field="*{eventImageList}" class="multi file-upload-input with-preview" multiple>
+                                            <input type="file" name="eventImageList" class="multi file-upload-input with-preview" multiple>
 <!--                                            <input type="file" th:field="*{eventImageList}" name="files[]" class="multi file-upload-input with-preview" multiple maxlength="3">-->
                                             <span class="file-upload-text"><i class="la la-upload mr-2"></i>파일을 선택하세요.</span>
                                         </div>
